+++ app.vue
@@ -4,7 +4,8 @@
 </template>
 
 <style>
-body {
+html {
+  scroll-behavior: smooth;  
   font-family: 'Baloo 2', sans-serif;
 }
 </style>