+++ components/auth/SignInForm.vue
@@ -1,7 +1,8 @@
 <template>
   <div class="mx-auto w-1/2 mt-10">
+    <img src="~/assets/objlab_logo.png" alt="objlab" class="w-56 cursor-pointer" @click="$router.push('/')">
     <h1 class="text-3xl font-bold leading-normal bg-gradient-to-r from-[#94543d] to-[#e2ad74] to-50% bg-clip-text text-transparent">Ravi de vous revoir !</h1>
-
+  
     <form @submit.prevent="handleUserSignin" class="mt-5 space-y-5">
       <div class="space-y-2">
         <label for="email">Email</label>